[fragment] Merge during download for `-N`, and refactor `hls`/`dash` (#364)
